Seperate random and demo data into individual seeders
