+++ app/seeders/random_data/board_seeder.rb
@@ -1,3 +1,4 @@
+#-- encoding: UTF-8
 #-- copyright
 # OpenProject is a project management system.
 # Copyright (C) 2012-2015 the OpenProject Foundation (OPF)
@@ -24,11 +25,34 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #
 # See doc/COPYRIGHT.rdoc for more details.
-#++
+module RandomData
+  class BoardSeeder
+    def self.seed!(project)
+      user = User.admin.first
 
-namespace 'demo_data' do
-  desc 'Seed the database with usefull demo data for onboarding users'
-  task :seed do
-    puts 'Seeded the database.'
+      puts ''
+      print ' ↳ Creating forum board with posts'
+
+      board = Board.create! project: project,
+                            name: I18n.t('seeders.demo_data.board.name'),
+                            description: I18n.t('seeders.demo_data.board.description')
+
+      rand(30).times do
+        print '.'
+        message = Message.create board: board,
+                                 author: user,
+                                 subject: Faker::Lorem.words(5).join(' '),
+                                 content: Faker::Lorem.paragraph(5, true, 3)
+
+        rand(5).times do
+          print '.'
+          Message.create board: board,
+                         author: user,
+                         subject: message.subject,
+                         content: Faker::Lorem.paragraph(5, true, 3),
+                         parent: message
+        end
+      end
+    end
   end
 end
