+++ app/seeders/random_data_seeder.rb
@@ -25,21 +25,30 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #
 # See doc/COPYRIGHT.rdoc for more details.
-class DemoDataSeeder
+class RandomDataSeeder
   def self.seed!
+    puts ' ########################################################'
+    puts ' #  WARNING: THIS DELETES ANY DEMO DATA THAT WAS SEEDED #'
+    puts ' ########################################################'
+
     project = DemoData::ProjectSeeder.seed!
 
     DemoData::CustomFieldSeeder.seed!(project)
-    DemoData::WikiSeeder.seed!(project)
-    DemoData::WorkPackageSeeder.new(project).seed!
+    RandomData::BoardSeeder.seed!(project)
+    RandomData::NewsSeeder.seed!(project)
+    RandomData::WikiSeeder.seed!(project)
+    RandomData::WorkPackageSeeder.new(project).seed!
 
     puts "\n\n"
-    puts ' ###############################'
-    puts ' #  Demo data seeding....done  #'
-    puts ' ###############################'
-    puts " #  %02d %-22s  #" % [1, 'project created.']
-    puts " #  %02d %-22s  #" % [WorkPackage.where(project_id: project.id).count, 'issues created.']
-    puts " #  %02d %-22s  #" % [WikiContent.joins(page: [:wiki]).where('wikis.project_id = ?', project.id).count, 'wiki created.']
-    puts " ###############################\n\n"
+    puts ' ################################'
+    puts ' #    Random seeding....done    #'
+    puts ' ################################'
+    puts " #  %02d %-23s  #" % [WorkPackage.where(project_id: project.id).count, 'issues created.']
+    puts " #  %02d %-23s  #" % [Message.joins(:board).where(boards: { project_id: project.id }).count, 'messages created.']
+    puts " #  %02d %-23s  #" % [News.where(project_id: project.id).count, 'news created.']
+    puts " #  %02d %-23s  #" % [WikiContent.joins(page: [:wiki]).where('wikis.project_id = ?', project.id).count, 'wiki contents created.']
+    puts " #  %02d %-23s  #" % [TimeEntry.where(project_id: project.id).count, 'time entries created.']
+    puts " #  %02d %-23s  #" % [Changeset.joins(:repository).where(repositories: { project_id: project.id }).count, 'changesets created.']
+    puts " ################################\n\n"
   end
 end
