+++ db/production.rb
@@ -1,0 +1,1 @@
+DemoDataSeeder.seed!
