+++ lib/tasks/random_data.rake
@@ -1,4 +1,3 @@
-#-- encoding: UTF-8
 #-- copyright
 # OpenProject is a project management system.
 # Copyright (C) 2012-2015 the OpenProject Foundation (OPF)
@@ -25,34 +24,24 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #
 # See doc/COPYRIGHT.rdoc for more details.
-module DemoData
-  class BoardSeeder
-    def self.seed!(project)
-      user = User.admin.first
+#++
 
-      puts ''
-      print ' ↳ Creating forum board with posts'
+namespace :random_data do
+  desc 'seeds the data base wth random data'
+  task seed: :environment do
 
-      board = Board.create! project: project,
-                            name: I18n.t('seeders.demo_data.board.name'),
-                            description: I18n.t('seeders.demo_data.board.description')
+    puts '*** Seeding basic data'
+    BasicDataSeeder.seed!
 
-      rand(30).times do
-        print '.'
-        message = Message.create board: board,
-                                 author: user,
-                                 subject: Faker::Lorem.words(5).join(' '),
-                                 content: Faker::Lorem.paragraph(5, true, 3)
+    puts '*** Seeding admin user'
+    AdminUserSeeder.seed!
 
-        rand(5).times do
-          print '.'
-          Message.create board: board,
-                         author: user,
-                         subject: message.subject,
-                         content: Faker::Lorem.paragraph(5, true, 3),
-                         parent: message
-        end
-      end
+    puts '*** Seeding demo data'
+    RandomDataSeeder.seed!
+
+    ::Rails::Engine.subclasses.map(&:instance).each do |engine|
+      puts "*** Loading #{engine.engine_name} seed data"
+      engine.load_seed
     end
   end
 end
